Program Core
These are the main files around which the backend of the Queueing system will operate. Commands that I have successfully implemented on both the server and the client end are: createroom, joinroom, leaveroom, and enqueue/dequeue. The transmission of these commands between client and server has been programmed as well as the dynamic creation and deletion of individual personalized room objects as requested with a simple but to be improved privileges system.